requirement.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
defusedxml==0.7.1
Django==6.0.1
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
httpie==3.2.4
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.7.0
Pygments==2.19.2
PyJWT==2.10.1
PySocks==1.7.1
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
setuptools==80.9.0
sqlparse==0.5.5
urllib3==2.6.3
